+++ editor/editor-core/src/__tests__/unit/plugins/card/paste.ts
@@ -51,7 +51,6 @@ describe('card', () => {
           cards: [],
           requests: [
             {
-              analyticsAction: 'pasted',
               url: 'https://www.atlassian.com/s/7xr7xdqto7trhvr/Media%20picker.sketch?dl=0',
               pos: 1,
               appearance: 'inline',