+++ editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/editor-mobile-bridge
 
+## 183.0.11
+
+### Patch Changes
+
+- Updated dependencies
+
+## 183.0.10
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 183.0.9
 
 ### Patch Changes